backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# AI and Data
pandasai==0.11.0
google-generativeai==0.3.2
pandas==2.1.3
openpyxl==3.1.2
plotly==5.18.0

# Configuration
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6
